+++ restaurant-siglo-xxi/src/main/java/com/restaurant/siglo/xxi/repository/ProductosPedidoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.restaurant.siglo.xxi.clases.ProductosPedido;
 
-public interface ProductosPedidoRepository extends JpaRepository<ProductosPedido, Integer>{
+public interface ProductosPedidoRepository extends JpaRepository<ProductosPedido, Long>{
 
 }
