+++ restaurant-siglo-xxi/src/main/java/com/restaurant/siglo/xxi/repository/ProductosRecetaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.restaurant.siglo.xxi.clases.ProductosReceta;
 
-public interface ProductosRecetaRepository extends JpaRepository<ProductosReceta, Integer> {
+public interface ProductosRecetaRepository extends JpaRepository<ProductosReceta, Long> {
 
 }
